requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
cachelib==0.12.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
distro==1.8.0
Flask==3.0.2
Flask-Limiter==3.5.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.6
limits==3.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
openai==1.3.7
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.0
priority==2.0.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
Quart==0.19.5
redis==5.0.3
requests==2.31.0
rich==13.7.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
